added static casts to help compilation

+++ src/ElectroMagnSolver/MA_SolverRZ_norm.cpp
@@ -43,8 +43,8 @@ void MA_SolverRZ_norm::operator() ( ElectroMagn* fields )
         for (unsigned int j=isYmin*3 ; j<nr_p ; j++) {
             (*ElRZ)(i,j) += -dt*(*JlRZ)(i,j)
                 +                 dt/((j_glob+j)*dr)* ((j+0.5)*(*BtRZ)(i,j+1) - (j-0.5)*(*BtRZ)(i,j) )
-                +                 Icpx*dt*imode/((j_glob+j)*dr)*(*BrRZ)(i,j);
-				if ((*ElRZ)(i,j)!= 0)
+                +                 Icpx*dt*static_cast<double>(imode)/((j_glob+j)*dr)*(*BrRZ)(i,j);
+				if ((*ElRZ)(i,j)!= 0.0)
 				{MESSAGE("ElRZ");				
 				MESSAGE(i);	
 				MESSAGE(j);
@@ -59,8 +59,8 @@ void MA_SolverRZ_norm::operator() ( ElectroMagn* fields )
         for (unsigned int j=0 ; j<nr_d ; j++) {
             (*ErRZ)(i,j) += -dt*(*JrRZ)(i,j)
                 -                  dt_ov_dl * ( (*BtRZ)(i+1,j) - (*BtRZ)(i,j) )
-                -                  Icpx*dt*imode/((j_glob+j+0.5)*dr)* (*BlRZ)(i,j);
-				if ((*ErRZ)(i,j)!= 0){
+                -                  Icpx*dt*static_cast<double>(imode)/((j_glob+j+0.5)*dr)* (*BlRZ)(i,j);
+				if ((*ErRZ)(i,j)!= 0.0){
 				MESSAGE("ErRZ");				
 				MESSAGE(i);
 				MESSAGE(j);	
@@ -76,7 +76,7 @@ void MA_SolverRZ_norm::operator() ( ElectroMagn* fields )
             (*EtRZ)(i,j) += -dt*(*JtRZ)(i,j)
                 +                  dt_ov_dl * ( (*BrRZ)(i+1,j) - (*BrRZ)(i,j) )
                 -                  dt_ov_dr * ( (*BlRZ)(i,j+1) - (*BlRZ)(i,j) );        
-				if ((*EtRZ)(i,j)!= 0){
+				if ((*EtRZ)(i,j)!= 0.0){
 				MESSAGE("EtRZ ");
 				MESSAGE(i);
 				MESSAGE(j);
